requirements.txt
# Amazon MQ Demo - Python Dependencies
# =====================================
# Install with: pip install -r requirements.txt

# Flask - Lightweight web framework for creating REST API
# We use Flask because it's simple and perfect for demos
flask==3.0.0

# Pika - Official RabbitMQ/AMQP client library for Python
# This connects our app to Amazon MQ (RabbitMQ)
pika==1.3.2

# Python-dotenv - Load environment variables from .env file
# Keeps sensitive credentials (passwords, URLs) out of code
python-dotenv==1.0.0

# Flask-CORS - Enable Cross-Origin Resource Sharing
# Allows Postman, browsers, and other tools to call our API
flask-cors==4.0.0
